# apis/gateway-admin-api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gateway Admin API",
    "description": "## Gateway Administration API\n\nOIDC-protected proxy to webMethods API Gateway administration endpoints.\n\n### Purpose\n\nThis API provides secure access to Gateway administration functions through the Control-Plane API, eliminating the need for direct Basic Auth access to the Gateway admin port (5555).\n\n### Authentication\n\nAll endpoints require a valid JWT token from Keycloak (realm: `stoa`).\n- Users with `cpi-admin` role can access all operations\n- Token is forwarded to Gateway for audit trail\n\n### Backend\n\nProxy to: `http://apigateway.stoa-system.svc.cluster.local:5555/rest/apigateway`\n\n### Features\n\n- **API Management**: List, activate, deactivate APIs\n- **Application Management**: Manage OAuth applications\n- **Scope Management**: Configure OAuth scopes\n- **OIDC Configuration**: Set up External Authorization Servers\n",
    "contact": {
      "name": "CAB Ingenierie",
      "email": "admin@cab-i.com"
    },
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.gostoa.dev/v1/gateway",
      "description": "Production - via Control-Plane API"
    }
  ],
  "paths": {
    "/apis": {
      "get": {
        "tags": ["APIs"],
        "summary": "List Gateway APIs",
        "description": "List all APIs registered in the Gateway",
        "operationId": "list_gateway_apis",
        "responses": {
          "200": {
            "description": "List of APIs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayAPIList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing JWT token"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [{"BearerAuth": []}]
      }
    },
    "/apis/{api_id}": {
      "get": {
        "tags": ["APIs"],
        "summary": "Get Gateway API",
        "description": "Get details of a specific API",
        "operationId": "get_gateway_api",
        "parameters": [
          {
            "name": "api_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "description": "Gateway API ID"
          }
        ],
        "responses": {
          "200": {
            "description": "API details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayAPI"
                }
              }
            }
          },
          "404": {
            "description": "API not found"
          }
        },
        "security": [{"BearerAuth": []}]
      }
    },
...
  },
  "components": {
    "schemas": {
      "GatewayAPI": {
        "type": "object",
        "properties": {
          "id": {"type": "string", "description": "Gateway API ID"},
          "apiName": {"type": "string"},
          "apiVersion": {"type": "string"},
          "apiDescription": {"type": "string"},
          "isActive": {"type": "boolean"},
          "type": {"type": "string", "enum": ["REST", "SOAP", "OData"]},
          "createdDate": {"type": "string", "format": "date-time"},
          "lastModified": {"type": "string", "format": "date-time"}
        }
      },
      "GatewayAPIList": {
        "type": "object",
        "properties": {
          "apiResponse": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/GatewayAPI"}
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token from Keycloak (realm: stoa)"
      }
    }
  },
...
}